requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
marshmallow==3.19.0
pytest==7.3.1
pytest-cov==4.1.0
Werkzeug==2.3.7